Improve last commit.
(cherry picked from commit 113a168b792cd367822ec04cdc2ef32facd28efa)

The `noconflict()` function checks if there's no conflicting instruction
between the current instruction and the given `ref` instruction. Also,
it avoids fusing loads if there are multiple references of the given
`ref`. The last check is performed in the presence of the `noload`
parameter. Since the `noconflict()`, which is added in the previous
patch, checks conflicts for the same `ref` as the call before, there is
no need to perform these checks again, so the corresponding parameter is
adjusted.

Sergey Kaplun:
* added the description for the problem

Part of tarantool/tarantool#10709

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>