cmake: run tests with Valgrind
This patch enables running tests with Valgrind. There is a
`VALGRIND_OPTS` variable [1] that we can set -- it makes the usage of
Valgrind more flexible -- we can define any necessary flags in the
command line (not at the building stage). By default, the suppression
files are set to <src/lj.supp> (original suppression file in LuaJIT) and
an additional one <src/lj_extra.supp> (maintained by us).

Also, this patch disables the following tests when running with Valgrind
due to failures:

The <tarantool-tests/lj-512-profiler-hook-finalizers.test.lua> test is
disabled due to its time sensitivity (it is not run the expected amount
of time with Valgrind).

These tests from the tarantool-tests suite are disabled due to
tarantool/tarantool#10803:
  - lj-726-profile-flush-close.test.lua
  - profilers/gh-5688-tool-cli-flag.test.lua
  - profilers/gh-7264-add-proto-trace-sysprof-default.test.lua
  - profilers/misclib-sysprof-lapi.test.lua

Timed out due to running under Valgrind:
  - tarantool-c-tests/gh-8594-sysprof-ffunc-crash.c_test
  - tarantool-tests/gh-7745-oom-on-trace.test.lua
  - tarantool-tests/lj-1034-tabov-error-frame.test.lua

[1]: https://valgrind.org/docs/manual/manual-core.html#manual-core.defopts

Part of tarantool/tarantool#3705

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>