Fix compilation of getmetatable() for UDTYPE_IO_FILE
The patch fixes a problem with recording `getmetatable()`
for an I/O object: recording the `getmetatable` call with file
descriptors represented by userdata object `UDTYPE_IO_FILE`
(like `io.stdout`) leads to violation of assertion in
`rec_check_slots`.

Note, the problem was fixed upstream in different manner, see
commit 5141cbc20c43921778ff36be541c15888bee8ee3
("Fix compiliation of getmetatable() for UDTYPE_IO_FILE.").
Note, the specialization omits the check of `__metatable` field
precense and the check for the metatable itself. So, if we change
the metatable on the object after the trace is compiled, the trace
becomes invalid. The proposed test demonstrates these cases as
well.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>