cmake: fix build for Alpine
Since Alpine uses musl [1] as its C standard library, the build for it
failed after the commit af0f59da76292f30ff75be1ab01458d47b226995 ("test:
fix LuaJIT-tests for old libc version"), since `GetLibCVersion()` raises
an error. This patch adds the check of the ID in the
</etc/os-release> [2][3] of the Linux distribution and avoids setting
the glibc version if the distro is "alpine".

[1]: https://wiki.alpinelinux.org/wiki/Musl
[2]: https://www.linux.org/docs/man5/os-release.html
[3]: https://www.freedesktop.org/software/systemd/man/latest/os-release.html

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>