sysprof: introduce specific errors and default mode
sysprof has a number of options and with any incorrect option it
returns `false` and error message "profiler misuse". This message
discourages sysprof users and makes using sysprof more
complicated. The patch sets the default profiling mode ("D", that
shows only virtual machine state counters) if it was not passed
and adds details to the error message with possible reasons of
misuse. Also, it fixes the bug when the error isn't raised if the
given argument type is incorrect. Also, the profiler can start if
it has extra arguments given. They are just ignored now.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>